code refactoring: fetch search data
